Add App Transport Security Exceptions to Info.plist
iOS 9 has this new feature that requires that all NSURLSessionManager connects use https and otherwise blocks the connection. I tried to set the NOAA connection urls to use https but they didn't work due to issues on the server. I've added exceptions to the Info.plist to allow getting the xml data from the server and the images from another server. This might stop working at some point but it works now.